requirements.txt
anyascii==0.3.1
asgiref==3.4.1
asttokens==2.0.5
backcall==0.2.0
beautifulsoup4==4.9.3
black==22.1.0
cachetools==5.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.0.4
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
dj-config-url==0.1.1
Django==3.2.9
django-admin-tools==0.9.2
django-dbbackup==3.3.0
django-extensions==3.1.5
django-filebrowser==4.0.1
django-filter==21.1
django-grappelli==3.0.3
django-import-export==2.7.1
django-jet-reboot==1.3.1
django-modelcluster==6.0
django-ninja==0.17.0
django-permissionedforms==0.1
django-taggit==2.1.0
django-tinymce==3.4.0
django-treebeard==4.5.1
djangorestframework==3.12.4
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
executing==0.8.2
feedparser==6.0.10
google-api-core==2.8.1
google-api-python-client==2.49.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.2
html5lib==1.1
httplib2==0.20.4
idna==3.3
ipython==8.0.1
jedi==0.18.1
l18n==2021.3
MarkupPy==1.14
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
odfpy==1.4.1
openpyxl==3.0.9
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==2.5.1
prompt-toolkit==3.0.28
protobuf==3.20.1
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.0
pydotplus==2.0.2
Pygments==2.11.2
pyparsing==3.0.9
python-decouple==3.5
pytz==2021.3
PyYAML==6.0
requests==2.27.1
rsa==4.8
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.2.0
tablib==3.2.0
telepath==0.2
tomli==2.0.1
traitlets==5.1.1
typing-extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.9
wagtail==3.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.2
Willow==1.4.1
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
